add/delete/change reaction implemented in one api

+++ src/main/java/com/sparta/doblock/reaction/service/ReactionService.java
@@ -5,10 +5,13 @@ import com.sparta.doblock.feed.repository.FeedRepository;
 import com.sparta.doblock.member.entity.MemberDetailsImpl;
 import com.sparta.doblock.reaction.dto.request.ReactionRequestDto;
 import com.sparta.doblock.reaction.entity.Reaction;
+import com.sparta.doblock.reaction.entity.ReactionType;
 import com.sparta.doblock.reaction.repository.ReactionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -21,12 +24,14 @@ public class ReactionService {
                 () -> new NullPointerException("해당 피드가 없습니다")
         );
 
+        reactionRequestDto.capitalize();
+
         if (! reactionRepository.existsByFeedAndMember(feed, memberDetails.getMember())) {
             // add reaction
             Reaction reaction = Reaction.builder()
                     .feed(feed)
                     .member(memberDetails.getMember())
-                    .reactionType(reactionRequestDto.getReactionType())
+                    .reactionType(ReactionType.valueOf(reactionRequestDto.getType()))
                     .build();
 
             reactionRepository.save(reaction);
@@ -36,9 +41,23 @@ public class ReactionService {
             Reaction reaction = reactionRepository.findByFeedAndMember(feed, memberDetails.getMember()).orElseThrow(
                     () -> new NullPointerException()
             );
-            reactionRepository.delete(reaction);
+            boolean delete = false;
+            if (Objects.isNull(reactionRequestDto.getType())) {
+                delete = true;
+            } else if (Objects.isNull(ReactionType.valueOf(reactionRequestDto.getType()))) {
+                delete = true;
+            } else if (ReactionType.valueOf(reactionRequestDto.getType()).equals(reaction.getReactionType())) {
+                delete = true;
+            }
 
-            return ResponseEntity.ok("성공적으로 리액션을 삭제하였습니다");
+            if (! delete) {
+                // change reaction type
+                reaction.update(ReactionType.valueOf(reactionRequestDto.getType()));
+                return ResponseEntity.ok("성공적으로 리액션을 수정하였습니다");
+            } else {
+                reactionRepository.delete(reaction);
+                return ResponseEntity.ok("성공적으로 리액션을 삭제하였습니다");
+            }
         }
     }
 }
